requirements.txt
langchain
langchain-openai
langchain-chroma
pypdf
python-docx
python-dotenv
langchain_community
flask
werkzeug